refactor meta schedule testing utils
